Split large `cfg_if!` into multiple modules
This commit splits the large `cfg_if!` which defines the
`native_libraries` functions in the gimli module into multiple modules.
This refactoring should hopefully make it a bit easier to grok what's
going on and avoid storing everything in a very large macro invocation.
